+++ src/ndv/__init__.py
@@ -8,9 +8,9 @@ except PackageNotFoundError:
     __version__ = "uninstalled"
 
 from . import data
-from ._views import run_app
 from .controllers import ArrayViewer
 from .models import DataWrapper
 from .util import imshow
+from .views import run_app
 
 __all__ = ["ArrayViewer", "DataWrapper", "data", "imshow", "run_app"]
