+++ src/ndv/controllers/_channel_controller.py
@@ -8,9 +8,9 @@ if TYPE_CHECKING:
 
     import numpy as np
 
-    from ndv._views.bases import LutView
-    from ndv._views.bases.graphics._canvas_elements import ImageHandle
     from ndv.models._lut_model import LUTModel
+    from ndv.views.bases import LutView
+    from ndv.views.bases.graphics._canvas_elements import ImageHandle
 
     LutKey = int | None
