+++ src/ndv/util.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, overload
 
-from ndv._views._app import run_app
 from ndv.controllers import ArrayViewer
+from ndv.views._app import run_app
 
 if TYPE_CHECKING:
     from typing import Any, Unpack
