+++ src/ndv/v1/_old_viewer.py
@@ -12,7 +12,7 @@ from qtpy.QtWidgets import QHBoxLayout, QLabel, QPushButton, QVBoxLayout, QWidge
 from superqt import QCollapsible, QElidingLabel, QIconifyIcon, ensure_main_thread
 from superqt.utils import qthrottled, signals_blocked
 
-from ndv._views import get_array_canvas_class
+from ndv.views import get_array_canvas_class
 
 from ._old_data_wrapper import DataWrapper
 from ._qt._components import (
@@ -33,8 +33,8 @@ if TYPE_CHECKING:
     from qtpy.QtCore import QObject
     from qtpy.QtGui import QCloseEvent, QKeyEvent
 
-    from ndv._views.bases.graphics._canvas import ArrayCanvas
-    from ndv._views.bases.graphics._canvas_elements import (
+    from ndv.views.bases.graphics._canvas import ArrayCanvas
+    from ndv.views.bases.graphics._canvas_elements import (
         CanvasElement,
         ImageHandle,
         RoiHandle,
