+++ src/ndv/views/_app.py
@@ -18,8 +18,8 @@ if TYPE_CHECKING:
 
     from IPython.core.interactiveshell import InteractiveShell
 
-    from ndv._views.bases import ArrayCanvas, ArrayView, HistogramCanvas
-    from ndv._views.bases.graphics._mouseable import Mouseable
+    from ndv.views.bases import ArrayCanvas, ArrayView, HistogramCanvas
+    from ndv.views.bases.graphics._mouseable import Mouseable
 
 
 GUI_ENV_VAR = "NDV_GUI_FRONTEND"
@@ -329,7 +329,7 @@ class VispyProvider(CanvasProvider):
     def array_canvas_class() -> type[ArrayCanvas]:
         from vispy.app import use_app
 
-        from ndv._views._vispy._array_canvas import VispyArrayCanvas
+        from ndv.views._vispy._array_canvas import VispyArrayCanvas
 
         # these may not be necessary, since we likely have already called
         # create_app by this point and vispy will autodetect that.
@@ -346,7 +346,7 @@ class VispyProvider(CanvasProvider):
 
     @staticmethod
     def histogram_canvas_class() -> type[HistogramCanvas]:
-        from ndv._views._vispy._histogram import VispyHistogramCanvas
+        from ndv.views._vispy._histogram import VispyHistogramCanvas
 
         return VispyHistogramCanvas
 
@@ -362,7 +362,7 @@ class PygfxProvider(CanvasProvider):
 
     @staticmethod
     def array_canvas_class() -> type[ArrayCanvas]:
-        from ndv._views._pygfx._array_canvas import GfxArrayCanvas
+        from ndv.views._pygfx._array_canvas import GfxArrayCanvas
 
         return GfxArrayCanvas
