+++ src/ndv/views/_qt/spin.gif
