+++ src/ndv/views/_wx/_labeled_slider.py
