+++ src/ndv/views/_wx/range_slider.py
