+++ src/ndv/views/bases/graphics/_mouseable.py
